[vault] Remove everything except unit helpers. This removes the gittin dependency. Contributes to JB#36999
The git-related features are no longer required, so remove them and
avoid the gittin dependency. Keep unit.hpp/cpp in the devel library
as it is still required to build backup/restore units.

Also build with qmake and rename unit.hpp to unit.h for consistency
with other Sailfish middleware libraries.